+++ BankingApp/banking_app/src/main/java/com/project/banking/dao/impl/AccountServiceDAOImpl.java
@@ -34,7 +34,7 @@ public class AccountServiceDAOImpl implements AccountServiceDAO {
 				Account_Details account=new Account_Details();
 				account.setA_id(a_id);
 				account.setC_id(c_id);
-				account.setBalance(resultSet.getInt("balance"));
+				account.setBalance(resultSet.getFloat("balance"));
 				accountList.add(account);
 			}
 		} catch (ClassNotFoundException | SQLException e) {
@@ -58,7 +58,7 @@ public class AccountServiceDAOImpl implements AccountServiceDAO {
 				account.setA_id(resultSet.getInt("a_id"));
 				account.setC_id(c_id);
 				account.setA_type(resultSet.getString("a_type"));
-				account.setBalance(resultSet.getInt("balance"));
+				account.setBalance(resultSet.getFloat("balance"));
 				accountList.add(account);
 			}
 		} catch (ClassNotFoundException | SQLException e) {
@@ -69,14 +69,14 @@ public class AccountServiceDAOImpl implements AccountServiceDAO {
 	}
 
 	@Override
-	public int conductTransaction(int a_id, String t_type, int amount) throws BusinessException {
+	public int conductTransaction(int a_id, String t_type, float amount) throws BusinessException {
 		int c=0;
 		try(Connection connection=PostgresConnection.getConnection()){
 			String sql = "insert into banking_app_schema.transaction_details(a_id,t_type,amount) values (?,?,?)" ;
 			PreparedStatement preparedStatement=connection.prepareStatement(sql);
 			preparedStatement.setInt(1, a_id);
 			preparedStatement.setString(2, t_type);
-			preparedStatement.setInt(3, amount);
+			preparedStatement.setFloat(3, amount);
 			c=preparedStatement.executeUpdate();
 		} catch (ClassNotFoundException | SQLException e) {
 			log.warn(e);
@@ -87,12 +87,12 @@ public class AccountServiceDAOImpl implements AccountServiceDAO {
 	}
 
 	@Override
-	public int updateAccount(int balance, int a_id) throws BusinessException {
+	public int updateAccount(float balance, int a_id) throws BusinessException {
 		int c=0;
 		try(Connection connection=PostgresConnection.getConnection()){
 			String sql = "update banking_app_schema.account_details set balance=? where a_id=?" ;
 			PreparedStatement preparedStatement=connection.prepareStatement(sql);
-			preparedStatement.setInt(1, balance);
+			preparedStatement.setFloat(1, balance);
 			preparedStatement.setInt(2, a_id);
 			c=preparedStatement.executeUpdate();
 		} catch (ClassNotFoundException | SQLException e) {
@@ -104,14 +104,14 @@ public class AccountServiceDAOImpl implements AccountServiceDAO {
 	}
 
 	@Override
-	public int createAccount(int c_id, String a_type, int balance) throws BusinessException {
+	public int createAccount(int c_id, String a_type, float balance) throws BusinessException {
 		int c=0;
 		try(Connection connection=PostgresConnection.getConnection()){
 			String sql = "insert into banking_app_schema.account_details(c_id,a_type,balance) values(?,?,?)";
 			PreparedStatement preparedStatement=connection.prepareStatement(sql);
 			preparedStatement.setInt(1, c_id);
 			preparedStatement.setString(2, a_type);
-			preparedStatement.setInt(3, balance);
+			preparedStatement.setFloat(3, balance);
 			c=preparedStatement.executeUpdate();
 		} catch (ClassNotFoundException | SQLException e) {
 			log.warn(e);
@@ -134,7 +134,7 @@ public class AccountServiceDAOImpl implements AccountServiceDAO {
 				Transaction_Details transaction=new Transaction_Details();
 				transaction.setA_id(a_id);
 				transaction.setT_type(resultSet.getString("t_type"));
-				transaction.setAmount(resultSet.getInt("amount"));
+				transaction.setAmount(resultSet.getFloat("amount"));
 				transaction.setDate(resultSet.getString("date"));
 				transactionList.add(transaction);
 			}
