+++ BankingApp/banking_app/src/main/java/com/project/banking/services/AccountService.java
@@ -10,9 +10,9 @@ public interface AccountService {
 
 	public List<Account_Details> getAccountBalance(int a_id, int c_id) throws BusinessException; 
 	public List<Account_Details> getCustomerAccounts(int c_id) throws BusinessException; 
-	public int conductTransaction(int a_id, String t_type, int amount) throws BusinessException;
-	public int updateAccount(int balance, int a_id) throws BusinessException;
-	public int createAccount(int c_id, String a_type, int balance) throws BusinessException;
+	public int conductTransaction(int a_id, String t_type, float amount) throws BusinessException;
+	public int updateAccount(float balance, int a_id) throws BusinessException;
+	public int createAccount(int c_id, String a_type, float balance) throws BusinessException;
 	public List<Transaction_Details> getTransLog(int a_id) throws BusinessException;
 	
 }
